Finish up stepper and some ui refactoring
